+++ project/src/components/app/app.tsx
@@ -1,12 +1,12 @@
 /* eslint-disable no-console */
 import { Route, Routes } from 'react-router-dom';
-import { AppRoute, AuthorizationStatus } from '../../const';
+import { AppRoute, isCheckedAuth } from '../../const';
 import PrivateRoute from '../private-route/private-route';
 import { useAppSelector } from '../../hooks';
 import MainScreen from '../../pages/main/main';
 import Film from '../../pages/film/film';
 import MyList from '../../pages/my-list/my-list';
-import SingIn from '../../pages/sing-in/sing-in';
+import SignIn from '../../pages/sign-in/sign-in';
 import AddReview from '../../pages/review/add-review';
 import Player from '../../pages/player/player';
 import NotFoundScreen from '../../pages/not-found-screen/not-found-screen';
@@ -15,9 +15,6 @@ import HistoryRouter from '../history-route/history-route';
 import browserHistory from '../../browser-history';
 import { getAuthorizationStatus } from '../../store/user-process/selectors';
 import { getLoadedDataStatus } from '../../store/films-data/selectors';
-
-const isCheckedAuth = (authorizationStatus: AuthorizationStatus): boolean =>
-  authorizationStatus === AuthorizationStatus.Unknown;
 
 function App(): JSX.Element {
   const authorizationStatus = useAppSelector(getAuthorizationStatus);
@@ -33,7 +30,7 @@ function App(): JSX.Element {
         <Route path={AppRoute.Main} element={<MainScreen />} />
         <Route path={AppRoute.Film} element={<Film />} />
         <Route path={AppRoute.AddReview} element={<AddReview />} />
-        <Route path={AppRoute.Login} element={<SingIn />} />
+        <Route path={AppRoute.Login} element={<SignIn />} />
         <Route path={AppRoute.Player} element={<Player />} />
         <Route
           path={AppRoute.MyList}
