+++ project/src/const.ts
@@ -49,3 +49,6 @@ export const HeaderCondition = {
   Main: true,
   Film: false,
 };
+export const isCheckedAuth = (
+  authorizationStatus: AuthorizationStatus
+): boolean => authorizationStatus === AuthorizationStatus.Unknown;
