+++ project/src/pages/main/main.tsx
@@ -1,30 +1,19 @@
 /* eslint-disable no-console */
 import Header from '../../components/header/header';
 import Footer from '../../components/footer/footer';
-import FilmList from '../../components/film-list/film-list';
-import GenreList from '../../components/genre-list/genre-list';
-import ShowMoreButton from '../../components/show-more/show-more';
+import Catalog from '../../components/catalog/catalog';
 import FilmCardDescp from '../../components/film-card/film-card-desc';
 import FilmCardPoster from '../../components/film-card/film-card-poster';
 import { useAppSelector } from '../../hooks';
-import { useState } from 'react';
 import { AuthorizationStatus, HeaderCondition } from '../../const';
 import { getAuthorizationStatus } from '../../store/user-process/selectors';
 import { getallFilmsList } from '../../store/films-data/selectors';
-import { getLoadMoreFilms } from '../../store/films-process/selectors';
 
 function MainScreen(): JSX.Element {
   const allFilmListFromState = useAppSelector(getallFilmsList);
-  const buttonConditionFromState = useAppSelector(getLoadMoreFilms);
   const authorizationStatus = useAppSelector(getAuthorizationStatus);
 
   const filmCard = allFilmListFromState[0];
-  const [isVisibleFilmButton, setVisibleFilmButton] = useState(
-    buttonConditionFromState
-  );
-  if (isVisibleFilmButton !== buttonConditionFromState) {
-    setVisibleFilmButton((prevState) => !prevState);
-  }
 
   return (
     <>
@@ -47,13 +36,7 @@ function MainScreen(): JSX.Element {
       </section>
 
       <div className="page-content">
-        <section className="catalog">
-          <h2 className="catalog__title visually-hidden">Catalog</h2>
-
-          <GenreList />
-          <FilmList />
-          {isVisibleFilmButton && <ShowMoreButton />}
-        </section>
+        <Catalog />
         <Footer />
       </div>
     </>
