+++ project/src/pages/sign-in/sign-in.tsx
@@ -1,0 +1,29 @@
+/* eslint-disable no-console */
+import Footer from '../../components/footer/footer';
+import Logo from '../../components/logo/logo';
+import SignInForm from './sign-in-form';
+import { AuthorizationStatus, AppRoute } from '../../const';
+import { Navigate } from 'react-router-dom';
+import { useAppSelector } from '../../hooks';
+import { getAuthorizationStatus } from '../../store/user-process/selectors';
+
+function SignIn(): JSX.Element {
+  const authorizationStatus = useAppSelector(getAuthorizationStatus);
+
+  return authorizationStatus === AuthorizationStatus.Auth ? (
+    <Navigate to={AppRoute.Main} />
+  ) : (
+    <div className="user-page">
+      <header className="page-header user-page__head">
+        <Logo />
+        <h1 className="page-title user-page__title">Sign in</h1>
+      </header>
+      <div className="sign-in user-page__content">
+        <SignInForm />
+      </div>
+      <Footer />
+    </div>
+  );
+}
+
+export default SignIn;
