--- project/src/pages/sing-in/sing-in.tsx
@@ -1,93 +1,0 @@
-/* eslint-disable no-console */
-import Footer from '../../components/footer/footer';
-import Logo from '../../components/logo/logo';
-import { useRef, FormEvent } from 'react';
-import { useAppDispatch } from '../../hooks';
-import { loginAction } from '../../store/api-actions';
-import { AuthData } from '../../types/auth-data';
-import { AuthorizationStatus, AppRoute } from '../../const';
-import { Navigate } from 'react-router-dom';
-import { useAppSelector } from '../../hooks';
-import { getAuthorizationStatus } from '../../store/user-process/selectors';
-
-function SingIn(): JSX.Element {
-  const authorizationStatus = useAppSelector(getAuthorizationStatus);
-
-  const loginRef = useRef<HTMLInputElement | null>(null);
-  const passwordRef = useRef<HTMLInputElement | null>(null);
-
-  const dispatch = useAppDispatch();
-
-  const onSubmit = (authData: AuthData) => {
-    dispatch(loginAction(authData));
-  };
-
-  const handleSubmit = (evt: FormEvent<HTMLFormElement>) => {
-    evt.preventDefault();
-
-    if (loginRef.current !== null && passwordRef.current !== null) {
-      onSubmit({
-        login: loginRef.current.value,
-        password: passwordRef.current.value,
-      });
-    }
-  };
-
-  return authorizationStatus === AuthorizationStatus.Auth ? (
-    <Navigate to={AppRoute.Main} />
-  ) : (
-    <div className="user-page">
-      <header className="page-header user-page__head">
-        <Logo />
-        <h1 className="page-title user-page__title">Sign in</h1>
-      </header>
-      <div className="sign-in user-page__content">
-        <form action="#" className="sign-in__form" onSubmit={handleSubmit}>
-          <div className="sign-in__fields">
-            <div className="sign-in__field">
-              <input
-                ref={loginRef}
-                className="sign-in__input"
-                type="email"
-                placeholder="Email address"
-                name="user-email"
-                id="user-email"
-              />
-
-              <label
-                className="sign-in__label visually-hidden"
-                htmlFor="user-email"
-              >
-                Email address
-              </label>
-            </div>
-            <div className="sign-in__field">
-              <input
-                ref={passwordRef}
-                className="sign-in__input"
-                type="password"
-                placeholder="Password"
-                name="user-password"
-                id="user-password"
-              />
-              <label
-                className="sign-in__label visually-hidden"
-                htmlFor="user-password"
-              >
-                Password
-              </label>
-            </div>
-          </div>
-          <div className="sign-in__submit">
-            <button className="sign-in__btn" type="submit">
-              Sign in
-            </button>
-          </div>
-        </form>
-      </div>
-      <Footer />
-    </div>
-  );
-}
-
-export default SingIn;
